% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sachintendulkar.R
\name{sachintendulkar}
\alias{sachintendulkar}
\title{An SRT Function}
\usage{
sachintendulkar(century = TRUE)
}
\arguments{
\item{What}{is Sachin's score? Defaults to TRUE.}
}
\description{
This function allows you to proclaim how incredible you think Sachin is.
}
\examples{
cat_function()
}
\keyword{Tendulkar}

